web/requirements.txt
Django==1.10.5
gunicorn==19.3.0
psycopg2
django-htmlmin
djangorestframework
django-filter
django-registration
